+++ packages/app/src/pages/Index.vue
@@ -4,25 +4,14 @@
       Loading...
     </div>
     <div v-else>
-      <ul>
-        <li v-for="spec of query.data.value?.app?.activeProject?.specs?.edges" :key="spec?.node?.absolute">
-          <!-- is there really no better way than the null coercion ?? -->
-          <button @click="setSpec(spec?.node ?? undefined)">
-            {{ spec?.node?.relative }}
-          </button>
-        </li>
-      </ul>
-
       <Specs 
         :gql="query.data.value?.app?.activeProject" 
-        :spec="store.spec"
       />
     </div>
   </div>
 </template>
 
 <script setup lang="ts">
-import { toJS } from 'mobx'
 import { gql, useQuery } from '@urql/vue'
 import Specs from './Specs.vue'
 import { store } from '../store'
